requirements.txt
apache-airflow-providers-amazon==1.3.0